# Conflicts:
#	composer.lock
#	src/Controller/DemandeController.php

+++ src/Service/RefArticleDataService.php
@@ -87,6 +87,7 @@ class RefArticleDataService
         ["title" => "Dernière entrée le", "name" => "lastStockEntry", "type" => "date"],
         ["title" => "Dernière sortie le", "name" => "lastStockExit", "type" => "date"],
         ["title" => "Inventaire à jour", "name" => "upToDateInventory", "type" => "booleen"],
+        ["title" => "Gestion quantité", "name" => "quantityType", "type" => "text"],
         ["title" => FiltreRef::FIXED_FIELD_VISIBILITY_GROUP, "name" => "visibilityGroups", "type" => "list multiple", "orderable" => true],
     ];
 
@@ -388,20 +389,22 @@ class RefArticleDataService
             $article->setVisible($isVisible);
         }
 
-        $managersIds = $data->get('managers');
-        $refArticle->getManagers()->clear();
-        if (!empty($managersIds)) {
-            $managers = is_string($managersIds)
-                ? explode(',', $managersIds)
-                : $managersIds;
-            foreach ($managers as $manager) {
-                $refArticle->addManager($userRepository->find($manager));
+        if($data->has('managers')) {
+            $managersIds = $data->get('managers');
+            $refArticle->getManagers()->clear();
+            if (!empty($managersIds)) {
+                $managers = is_string($managersIds)
+                    ? explode(',', $managersIds)
+                    : $managersIds;
+                foreach ($managers as $manager) {
+                    $refArticle->addManager($userRepository->find($manager));
+                }
             }
         }
 
         $typeId = $data->getInt('type');
         $type = $typeId ? $typeRepository->find($typeId) : null;
-        if ($type) {
+        if (!$refArticle->getId() && $type) {
             $refArticle->setType($type);
         }
 
@@ -409,18 +412,54 @@ class RefArticleDataService
             $refArticle->setLibelle($data->get('libelle'));
         }
 
-        $categoryId = $data->getInt('categorie');
-        $category = $categoryId ? $inventoryCategoryRepository->find($categoryId) : null;
+        if($data->has('categorie')) {
+            $categoryId = $data->getInt('categorie');
+            $refArticle->setCategory($inventoryCategoryRepository->find($categoryId));
+        }
 
-        $buyerId = $data->getInt('buyer');
-        $buyer = $buyerId ? $userRepository->find($buyerId) : null;
+        if($data->has('buyer')) {
+            $buyerId = $data->getInt('buyer');
+            $buyer = $buyerId ? $userRepository->find($buyerId) : null;
+            $refArticle->setBuyer($buyer);
+        }
 
-        $visibilityGroupId = $data->getInt('visibility-group');
-        $visibilityGroup = $visibilityGroupId ? $visibilityGroupRepository->find($visibilityGroupId) : null;
+        if($data->has('visibility-group')) {
+            $visibilityGroupId = $data->getInt('visibility-group');
+            $visibilityGroup = $visibilityGroupId ? $visibilityGroupRepository->find($visibilityGroupId) : null;
 
-        $isUrgent = $data->getBoolean('urgence');
+            $refArticle->setProperties(['visibilityGroup' => $visibilityGroup]);
+        }
 
-        $mobileSync= $data->getBoolean('mobileSync');
+        if ($data->has('limitWarning')) {
+            $limitWarning = $data->get('limitWarning');
+            $refArticle
+                ->setLimitWarning($limitWarning >= 0 ? $limitWarning : null);
+        }
+
+        if ($data->has('limitSecurity')) {
+            $limitSecurity = $data->get('limitSecurity');
+            $refArticle
+                ->setLimitSecurity($limitSecurity >= 0 ? $limitSecurity : null);
+        }
+
+        if ($data->has('prix')) {
+            $unitPrice = $data->get('prix');
+            $refArticle
+                ->setPrixUnitaire(($unitPrice !== null && $unitPrice >= 0) ? max(0, $unitPrice) : null);
+        }
+
+        if ($data->has('urgence')) {
+            $isUrgent = $data->getBoolean('urgence');
+            $emergencyQuantity = $data->get('emergencyQuantity');
+            $refArticle
+                ->setIsUrgent($isUrgent)
+                ->setUserThatTriggeredEmergency($isUrgent ? $user : null)
+                ->setEmergencyComment($isUrgent ? $data->get('emergencyComment') : '')
+                ->setEmergencyQuantity(($isUrgent && $emergencyQuantity >= 0) ? $emergencyQuantity : null);
+        }
+
+
+        $mobileSync = $data->getBoolean('mobileSync');
         if ($mobileSync) {
             $referenceArticleRepository = $entityManager->getRepository(ReferenceArticle::class);
             $syncCount = $referenceArticleRepository->count(['needsMobileSync' => true]);
@@ -433,24 +472,16 @@ class RefArticleDataService
         }
 
         $refArticle
-            ->setCategory($category)
             ->setReference($data->get('reference'))
-            ->setIsUrgent($isUrgent)
-            ->setUserThatTriggeredEmergency($isUrgent ? $user : null)
-            ->setEmergencyComment($isUrgent ? $data->get('emergencyComment') : '')
-            ->setEmergencyQuantity($isUrgent ? ($data->getInt('emergencyQuantity') >= 0) ? $data->getInt('emergencyQuantity') : null : null)
-            ->setPrixUnitaire(max(0, $data->get('prix')))
             ->setCommentaire($data->get('commentaire'))
             ->setNeedsMobileSync($mobileSync)
-            ->setBuyer($buyer)
-            ->setLimitWarning(($data->getInt('limitWarning') >= 0) ? $data->getInt('limitWarning') : null)
-            ->setLimitSecurity(($data->getInt('limitSecurity') >= 0) ? $data->getInt('limitSecurity') : null)
             ->setStockManagement($data->get('stockManagement'))
             ->setNdpCode($data->get('ndpCode'))
             ->setDangerousGoods($data->getBoolean('security'))
             ->setOnuCode($data->get('onuCode'))
             ->setProductClass($data->get('productClass'))
-            ->setProperties(['visibilityGroup' => $visibilityGroup]);
+            ->setEditedBy($user)
+            ->setEditedAt(new DateTime('now'));;
 
         if ($refArticle->getTypeQuantite() === ReferenceArticle::QUANTITY_TYPE_REFERENCE &&
             $refArticle->getQuantiteStock() > 0 &&
@@ -470,10 +501,6 @@ class RefArticleDataService
             $entityManager->persist($mvtStock);
         }
 
-        $refArticle
-            ->setEditedBy($user)
-            ->setEditedAt(new DateTime('now'));
-
         $entityManager->persist($refArticle);
         //modification ou création des champsLibres
         $this->freeFieldService->manageFreeFields($refArticle, $data->all(), $entityManager);
@@ -485,7 +512,7 @@ class RefArticleDataService
 
                 $refArticle->setImage($attachments[0]);
                 $fileBag->remove('image');
-            } elseif ($data->getBoolean('deletedImage')) {
+            } else if ($data->getBoolean('deletedImage')) {
                 $image = $refArticle->getImage();
                 if ($image) {
                     $this->attachmentService->deleteAttachment($image);
@@ -500,7 +527,7 @@ class RefArticleDataService
 
                 $refArticle->setSheet($attachments[0]);
                 $fileBag->remove('fileSheet');
-            } else {
+            } else if ($data->getBoolean('deletedImage')) {
                 $image = $refArticle->getSheet();
                 if ($image) {
                     $this->attachmentService->deleteAttachment($image);
@@ -561,7 +588,7 @@ class RefArticleDataService
             ]),
             "label" => $formatService->referenceArticle($refArticle,  "Non défini", true),
             "reference" => $refArticle->getReference() ?? "Non défini",
-            "quantityType" => $refArticle->getTypeQuantite() ?? "Non défini",
+            "quantityType" => $refArticle->getTypeQuantite() === ReferenceArticle::QUANTITY_TYPE_ARTICLE ? 'Article' : 'Référence',
             "type" => "<div class='d-flex align-items-center'><span class='dt-type-color mr-2' style='background-color: $typeColor;'></span>"
                 . $formatService->type($refArticle->getType())
                 . "</div>",
